+++ mini/dodge/scripts/characters/train.gd
@@ -1,5 +1,9 @@
-extends Area2D
+extends KinematicBody2D
 # A script for the train from the "Dodge!" minigame.
+
+
+# Emitted if the player crashes.
+signal crashed
 
 
 # Represents what tracks the train might be on.
@@ -16,12 +20,19 @@ var _unlock_controls: bool = false setget unlock
 # The mainloop: Moves the train up and down the tracks on player input.
 func _process( _delta ):
 	if _unlock_controls:
-		var velocity = 0.0
+		var velocity: Vector2 = Vector2.ZERO
 		if Input.is_action_pressed( "move_up" ):
-			velocity -= MOVEMENT_SPEED
+			velocity.y -= MOVEMENT_SPEED
 		if Input.is_action_pressed( "move_down" ):
-			velocity += MOVEMENT_SPEED
-		position.y += velocity * _delta
+			velocity.y += MOVEMENT_SPEED
+		var collision = move_and_collide( velocity * _delta )
+		if collision:
+			var body = collision.collider
+			# Only fail the minigame if the body has a signal (we use this to
+			# differentiate the cars from the boundaries of the road)
+			if body.has_signal( "crashed" ):
+				emit_signal( "crashed" )
+				body.unlock( false )
 
 
 # Unlocks the controls for the train.